{{ cookiecutter.repo_name }}/requirements.txt
# external requirements
Sphinx
awscli
python-dotenv>=0.5.1